requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
sqlalchemy
beautifulsoup4
python-jose[cryptography]